ref: Move samples to their own dir.
